Move HOC folder outside of components folder
