+++ imports/ui/components/forgot_password.jsx
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom';
 import { Alert, Container, Button, Form, Row, Col } from 'react-bootstrap';
 import { Accounts } from 'meteor/accounts-base';
 import { EMAIL_REGEX } from '/imports/constants/regex';
-import { AuthContext } from './hoc/AuthProvider';
+import { AuthContext } from '../hoc/AuthProvider';
 import { Redirect } from 'react-router-dom';
 
 class ForgotPassword extends React.Component {
