+++ imports/ui/components/navbar.jsx
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom';
 import { Navbar, Nav, Button, Container } from 'react-bootstrap';
 import { Meteor } from 'meteor/meteor';
 import { PropTypes } from 'prop-types';
-import { AuthContext } from './hoc/AuthProvider';
+import { AuthContext } from '../hoc/AuthProvider';
 
 class NavigationBar extends React.Component {
   constructor(props) {
